fix : edit logic of creation of new board
- change NewBoard
- add NewBoardMainPlaceholder
- delete ModalBoard
- rename default small background image
- fix other small things
