+++ src/components/MainPlaceholder/MainPlaceholder.jsx
@@ -1,56 +1,13 @@
-import ModalBoard from 'components/ModalBoard/ModalBoard';
+import NewBoardMainPlaceholder from 'components/ModalBoard/NewBoardMainPlaceholder';
 import { useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { addBoard } from 'redux/boards/operations';
-import { selectIsBoardsLoading } from 'redux/boards/selectors';
-import styles from './MainPlaceholder.module.css';
+import { useSelector } from 'react-redux';
 import { selectTheme } from 'redux/auth/selectors';
+import styles from './MainPlaceholder.module.css';
 
 const MainPlaceholder = () => {
   const theme = useSelector(selectTheme);
   const [isModalOpen, setIsModalOpen] = useState(false);
   const toggleModal = () => setIsModalOpen(state => !state);
-  const isBoardsLoading = useSelector(selectIsBoardsLoading);
-  const required = true;
-  const [icon, setIcon] = useState('icon-project');
-  const [background, setBackground] = useState(null);
-
-  const dispatch = useDispatch();
-
-  const handleSubmit = event => {
-    event.preventDefault();
-    dispatch(
-      addBoard({
-        title: event.target[0].value,
-        icon,
-        background,
-      })
-    ).then(() => {
-      if (!isBoardsLoading) {
-        toggleModal();
-        setIcon('icon-project');
-        setBackground(null);
-      }
-    });
-  };
-
-  const changeIcon = event => {
-    setIcon(event.target.value);
-  };
-  const changeBg = event => {
-    setBackground(event.target.value);
-  };
-
-  const modalProps = {
-    isModalOpen,
-    toggleModal,
-    handleSubmit,
-    changeBg,
-    changeIcon,
-    icon,
-    background,
-    required,
-  };
 
   return (
     <div className={styles.mainDashboardContainer}>
@@ -69,12 +26,9 @@ const MainPlaceholder = () => {
         board serves as a powerful tool to organize the workflow and ensure
         effective collaboration among team members.
       </p>
-
-      <ModalBoard
-        {...modalProps}
-        modalTitle="New board"
-        submitButtonText="Create"
-      />
+      {isModalOpen && (
+        <NewBoardMainPlaceholder setIsModalOpen={setIsModalOpen} />
+      )}
     </div>
   );
 };
