+++ src/constants/backgrounds.js
@@ -1,12 +1,12 @@
 import airBalloon from '../images/smallBg/airBalloon.png';
 import blue from '../images/smallBg/blue.png';
 import cappadocia from '../images/smallBg/cappadocia.png';
+import base from '../images/smallBg/default.png';
 import flowers from '../images/smallBg/flowers.png';
 import gorge from '../images/smallBg/gorge.png';
 import greens from '../images/smallBg/greens.png';
 import moon from '../images/smallBg/moon.png';
 import mountains from '../images/smallBg/mountains.png';
-import base from '../images/smallBg/null.png';
 import rocksAndSea from '../images/smallBg/rocksAndSea.png';
 import sea from '../images/smallBg/sea.png';
 import semiMoon from '../images/smallBg/semiMoon.png';
@@ -35,7 +35,7 @@ export const title = [
 ];
 
 export const backgrounds = [
-  { title: 'null', src: base },
+  { title: 'default', src: base },
   { title: 'flowers', src: flowers },
   { title: 'mountains', src: mountains },
   { title: 'tree', src: tree },
